impulse-notifier/requirements.txt
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
tvDatafeed @ git+https://github.com/rongardF/tvdatafeed.git
